Added better path information for failure states of oneOf and allOf. (#27)
Replaced use of path! macro in places where it was not necessary, since it was only a single-element path

Fixed path! macro to use absolute path to `smallvec::smallvec!` macro if the feature is enabled,
instead of relative paths which fails unless `smallvec!` is already in scope, such as in tests.

Fixes #26